requirements.txt
numpy
opencv-python
Pillow
streamlit